requirements.txt
Django
celery
python-dotenv
Pillow
redis
django-celery-results